requirements.txt
django
djangorestframework
django-cors-headers
django-filter
requests
pyyaml
drf-yasg